pyproject.toml
[project]
name = "ym-scheduler"
version = "0.1.0"
description = "Config-driven scheduling framework for Young Men's / Youth combined and separate activities"
authors = [{ name = "YOUR NAME", email = "you@example.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
  "pyyaml",
  "python-dateutil",
  "ics",
  "rich",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
python_files = ["test_*.py"]
